requirements.txt
requests
beautifulsoup4
schedule
mypy  # For static type checking
types-requests  # Type stubs for requests
types-beautifulsoup4  # Type stubs for beautifulsoup4
python-dotenv  # For loading environment variables from .env file